backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
stripe==11.4.0
google-genai==1.14.0
elevenlabs==2.3.0
Pillow==11.1.0
aiofiles==24.1.0
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
